--- etc/systemd/system/zenaida-cleanup.service.example
@@ -1,31 +1,0 @@
-# Zenaida cleanup inactive users and activation codes systemd service configuration.
-#
-# Copy and modify `zenaida-cleanup.service` file to your local systemd folder to enable the service:
-#
-#         mkdir -p /home/zenaida/.config/systemd/user/
-#         cd /home/zenaida/zenaida/
-#         cp etc/systemd/system/zenaida-cleanup.service.example /home/zenaida/.config/systemd/user/zenaida-cleanup.service
-#         systemctl --user enable zenaida-cleanup.service
-#
-#
-# To start Zenaida Poll service run this command:
-#
-#         systemctl --user start zenaida-cleanup.service
-#
-#
-# You can always check current situation with:
-#
-#         systemctl --user status zenaida-cleanup.service
-#
-
-[Unit]
-Description=ZenaidaCleanup
-After=network.target
-
-[Service]
-Type=simple
-WorkingDirectory=/home/zenaida/zenaida/
-ExecStart=/bin/sh -c "/home/zenaida/zenaida/venv/bin/python /home/zenaida/zenaida/src/manage.py clean_inactive_activations_and_accounts 1>>/home/zenaida/logs/cleanup 2>>/home/zenaida/logs/cleanup.err"
-
-[Install]
-WantedBy=multi-user.target
